refactor: 얼럿창을 UIViewController의 extension으로 추가 및 post 이미지네임 해시값으로 수정
